refactor: move prompt and context menu to components
