+++ src/lib.rs
@@ -30,6 +30,4 @@ pub use verso::Verso;
 /// Re-exporting Winit for the sake of convenience.
 pub use winit;
 /// Context
-pub mod context_menu;
-/// Prompt
-pub mod prompt;
+pub mod components;
